Libretro common update

+++ src/libretro-common/encodings/encoding_crc32.c
@@ -1,4 +1,4 @@
-/* Copyright  (C) 2010-2018 The RetroArch team
+/* Copyright  (C) 2010-2020 The RetroArch team
  *
  * ---------------------------------------------------------------------------------------
  * The following license statement only applies to this file (encoding_crc32.c).
@@ -117,13 +117,13 @@ uint32_t file_crc32(uint32_t crc, const char *path)
    if (!buf)
       goto error;
 
-   for(i = 0; i < CRC32_MAX_MB; i++)
+   for (i = 0; i < CRC32_MAX_MB; i++)
    {
-      int nread = filestream_read(file, buf, CRC32_BUFFER_SIZE);
-      if (nread < 0)
+      int64_t nread = filestream_read(file, buf, CRC32_BUFFER_SIZE);
+      if (nread < 0)		
          goto error;
 
-      crc = encoding_crc32(crc, buf, nread);
+      crc = encoding_crc32(crc, buf, (size_t)nread);
       if (filestream_eof(file))
          break;
    }
